pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wellnest"
version = "0.1.0"
description = "Health equity analytics for Chicago — ingestion, modeling, and community dashboards"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "ChiEAC", email = "engineering@chieac.org" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: GIS",
]

dependencies = [
    # -- core data stack --
    "polars>=1.39,<2",
    "duckdb>=1.5,<2",           # switched from pandas — never looking back
    "geopandas>=1.0,<2",
    "shapely>=2.0,<3",
    "h3>=4.1,<5",               # uber's hex grid; we use res 8 for Chicago tracts
    "psycopg2-binary>=2.9,<3",  # binary wheel so we skip the libpq build dance
    "sqlalchemy>=2.0,<3",
    "python-dotenv>=1.0,<2",
    "pydantic>=2.12,<3",
    "pydantic-settings>=2.7,<3",

    # -- orchestration --
    "dagster>=1.12,<2",
    "dagster-webserver>=1.12,<2",
    "dagster-postgres>=0.28,<1",

    # -- transformation --
    "dbt-core>=1.10,<1.11",     # 1.11 not published on PyPI yet for postgres adapter
    "dbt-postgres>=1.10,<1.11",

    # -- data quality --
    "soda-core>=3.5,<3.6",
    "soda-core-postgres>=3.5,<3.6",

    # -- ml --
    "scikit-learn>=1.6,<2",
    "xgboost>=3.2,<4",
    "lightgbm>=4.6,<5",
    "mlflow>=3.10,<4",

    # -- ai / rag --
    "openai>=2.26,<3",
    "langchain>=0.3,<0.4",
    "langchain-openai>=0.3,<0.4",
    "langchain-community>=0.3,<0.4",
    "faiss-cpu>=1.9,<2",        # gpu build is a pain on CI; cpu is fine for our index size

    # -- api --
    "fastapi>=0.135,<1",
    "uvicorn[standard]>=0.34,<1",
    "httpx>=0.28,<1",

    # -- dashboard --
    "streamlit>=1.55,<2",
    "plotly>=6.0,<7",
    "folium>=0.19,<1",
    "streamlit-folium>=0.25,<1",

    # -- reports --
    "weasyprint>=63,<70",       # needs system cairo — see docs/setup.md
    "fpdf2>=2.8,<3",

    # -- utilities --
    "requests>=2.32,<3",
    "tenacity>=9.0,<10",        # retry decorator for flaky CDC endpoints
    "tqdm>=4.67,<5",
    "structlog>=24.4,<26",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9",
    "pytest-cov>=6.0,<7",
    "pytest-asyncio>=0.25,<1",
    "ruff>=0.15,<1",
    "mypy>=1.15,<2",
    "pre-commit>=4.1,<5",
    "ipykernel>=6.29",          # jupyter support for scratch notebooks
]

[project.scripts]
wellnest-api = "api.main:run"

[tool.setuptools.packages.find]
include = [
    "ingestion*",
    "orchestration*",
    "ml*",
    "ai*",
    "api*",
    "dashboard*",
    "reports*",
    "tests*",
]

# --------------------------------------------------------------------------- #
# ruff
# --------------------------------------------------------------------------- #
[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "T20",  # flake8-print — catch stray print() calls before they hit prod
    "RUF",  # ruff-specific
]
ignore = [
    "E501",   # line length handled by formatter
    "B008",   # function call in default arg — FastAPI Depends() triggers this
    "RUF012", # mutable class attrs — pydantic models need these
]

# sklearn / numpy convention uses X, X_train, etc.; renaming would hurt readability
[tool.ruff.lint.per-file-ignores]
"ml/**/*.py" = ["N806", "N803"]
"orchestration/assets/ml_assets.py" = ["N806", "N803"]

[tool.ruff.lint.isort]
known-first-party = ["wellnest", "ingestion", "transformation", "ml", "ai", "api", "dashboard"]

[tool.ruff.format]
quote-style = "double"

# --------------------------------------------------------------------------- #
# mypy
# --------------------------------------------------------------------------- #
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false          # too noisy while prototyping
ignore_missing_imports = true          # half our geo stack has no stubs
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false

# --------------------------------------------------------------------------- #
# pytest
# --------------------------------------------------------------------------- #
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests that hit real databases or APIs",
    "integration: end-to-end pipeline tests",
]
asyncio_mode = "auto"

# --------------------------------------------------------------------------- #
# coverage
# --------------------------------------------------------------------------- #
# Dashboard and reports are Streamlit/CLI entrypoints not executed under pytest; including
# them drags overall % down and breaks CI. Orchestration is Dagster-only (omitted).
[tool.coverage.run]
source = ["ingestion", "transformation", "ml", "ai", "api"]
omit = ["tests/*", "scripts/*", "orchestration/*", "dashboard/**", "reports/**"]

[tool.coverage.report]
fail_under = 35
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__",
]
